Support retrieving multiple work orders by property references
The existing endpoint will continue to work mostly as before. However,
it will no longer return a 404 result if the property does not exist,
as this becomes confusing when dealing with multiple properties at a
time. This behaviour isn't relied on in the frontend as far as I'm
aware, so shouldn't affect the integration.

The endpoint can now take multiple &propertyreference query params,
and will query on all properties given.